+++ src/aks-preview/azext_aks_preview/_help.py
@@ -634,6 +634,9 @@ helps['aks create'] = f"""
         - name: --if-none-match
           type: string
           short-summary: Set to '*' to allow a new cluster to be created, but to prevent updating an existing cluster. Other values will be ignored.
+        - name: --enable-static-egress-gateway
+          type: bool
+          short-summary: Enable Static Egress Gateway addon to the cluster.
     examples:
         - name: Create a Kubernetes cluster with an existing SSH public key.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --ssh-key-value /path/to/publickey
@@ -1205,6 +1208,12 @@ helps['aks update'] = """
         - name: --disable-network-observability
           type: bool
           short-summary: Disable network observability on a cluster
+        - name: --enable-advanced-network-observability
+          type: bool
+          short-summary: Enable advanced network observability functionalities on a cluster.
+        - name: --disable-advanced-network-observability
+          type: bool
+          short-summary: Disable advanced network observability functionalities on a cluster
         - name: --enable-cost-analysis
           type: bool
           short-summary: Enable exporting Kubernetes Namespace and Deployment details to the Cost Analysis views in the Azure portal. For more information see aka.ms/aks/docs/cost-analysis.
@@ -1234,6 +1243,12 @@ helps['aks update'] = """
         - name: --bootstrap-container-registry-resource-id
           type: string
           short-summary: Configure container registry resource ID. Must use "Cache" as bootstrap artifact source.
+        - name: --enable-static-egress-gateway
+          type: bool
+          short-summary: Enable Static Egress Gateway addon to the cluster.
+        - name: --disable-static-egress-gateway
+          type: bool
+          short-summary: Disable Static Egress Gateway addon to the cluster.
     examples:
       - name: Reconcile the cluster back to its current state.
         text: az aks update -g MyResourceGroup -n MyManagedCluster
@@ -1829,6 +1844,9 @@ helps['aks nodepool add'] = """
         - name: --if-none-match
           type: string
           short-summary: Set to '*' to allow a new agentpool to be created, but to prevent updating an existing agentpool. Other values will be ignored.
+        - name: --gateway-prefix-size
+          type: int
+          short-summary: The size of Public IPPrefix attached to the Gateway-mode node pool. The node pool must be in Gateway mode.
     examples:
         - name: Create a nodepool in an existing AKS cluster with ephemeral os enabled.
           text: az aks nodepool add -g MyResourceGroup -n nodepool1 --cluster-name MyManagedCluster --node-osdisk-type Ephemeral --node-osdisk-size 48