+++ src/aks-preview/azext_aks_preview/custom.py
@@ -491,7 +491,7 @@ def aks_create(
     disable_public_fqdn=False,
     service_principal=None,
     client_secret=None,
-    enable_managed_identity=True,
+    enable_managed_identity=False,
     assign_identity=None,
     assign_kubelet_identity=None,
     enable_aad=False,
@@ -603,6 +603,7 @@ def aks_create(
     enable_cilium_dataplane=False,
     custom_ca_trust_certificates=None,
     enable_network_observability=None,
+    enable_advanced_network_observability=None,
     # nodepool
     crg_id=None,
     message_of_the_day=None,
@@ -646,6 +647,8 @@ def aks_create(
     cluster_service_load_balancer_health_probe_mode=None,
     if_match=None,
     if_none_match=None,
+    # Static Egress Gateway
+    enable_static_egress_gateway=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()
@@ -827,6 +830,8 @@ def aks_update(
     safeguards_excluded_ns=None,
     enable_network_observability=None,
     disable_network_observability=None,
+    enable_advanced_network_observability=None,
+    disable_advanced_network_observability=None,
     # metrics profile
     enable_cost_analysis=False,
     disable_cost_analysis=False,
@@ -846,6 +851,9 @@ def aks_update(
     cluster_service_load_balancer_health_probe_mode=None,
     if_match=None,
     if_none_match=None,
+    # Static Egress Gateway
+    enable_static_egress_gateway=False,
+    disable_static_egress_gateway=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()
@@ -1293,6 +1301,8 @@ def aks_agentpool_add(
     enable_vtpm=False,
     if_match=None,
     if_none_match=None,
+    # static egress gateway - gateway-mode pool
+    gateway_prefix_size=None,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()