+++ src/containerapp/azext_containerapp/_models.py
@@ -292,6 +292,20 @@ ContainerAppsJob = {
     "tags": None
 }
 
+SessionPool = {
+    "location": None,
+    "properties": {
+        "environmentId": None,
+        "poolManagementType": None,
+        "containerType": None,
+        "customContainerTemplate": None,
+        "secrets": None,
+        "dynamicPoolConfiguration": None,
+        "scaleConfiguration": None,
+        "sessionNetworkConfiguration": None
+    }
+}
+
 DaprComponentResiliency = {
     "properties": {
         "inboundPolicy": {
@@ -586,4 +600,14 @@ CustomDomainConfiguration = {
     "certificateValue": None,
     "certificatePassword": None,
     "certificateKeyVaultProperties": None
+}
+
+RuntimeJava = {
+    "enableMetrics": False
+}
+
+DotNetComponent = {
+    "properties": {
+        "componentType": "AspireDashboard"
+    }
 }