+++ src/k8s-runtime/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+1.0.0
+++++++
+* Use stable API version `2024-03-01` for ARM API
+
 1.0.0b2
 ++++++
 * Fix `TypeError: 'NoneType' object is not callable` error when deleting storage class